feat: 그리드 부분을 세가지 파일 재구성 및 그리드 바 컬러 체인지
원래 한 파일에 있었는데, 더 복잡해지는 거 같아서, 그리드의 화면 별로 파일을 재구성하였으며, 그리드 바에서 디테일 보기, 로고만 보기 버튼을 클릭시 컬러가 변경되게 제작하였습닏나.
